refactor: move code to src folder
