+++ src/tasks/delete-templates.js
